+++ lib/hexapdf/cli/merge.rb
@@ -181,7 +181,7 @@ module HexaPDF
         if rotation == :none
           page.delete(:Rotate)
         elsif rotation.kind_of?(Integer)
-          page[:Rotate] = ((page[:Rotate] || 0) + rotation) % 360
+          page.rotate(rotation)
         end
         page_tree.document.add(page)
         page_tree.add_page(page)