+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [HexaPDF::Type::FontType3] for Type 3 font support
 * [HexaPDF::Content::GraphicObject::Geom2D] for [Geom2D] object drawing support
 * [HexaPDF::Type::Page#orientation] for easy determination of page orientation
+* [HexaPDF::Type::Page#rotate] for rotating a page
 
 ### Changed